move fluid.layers.crf_decoding to static.nn.common
